MINOR: h1: Raise the chunk size limit up to (2^52 - 1)
The allowed chunk size was historically limited to 2GB to avoid risk of
overflow. This restriction is no longer necessary because the chunk size is
immediately stored into a 64bits integer after the parsing. Thus, it is now
possible to raise this limit. However to never fed possibly bogus values
from languages that use floats for their integers, we don't get more than 13
hexa-digit (2^52 - 1). 4 petabytes is probably enough !

This patch should fix the issue #1065. It may be backported as far as
2.1. For the 2.0, the legacy HTTP part must be reviewed. But there is
honestely no reason to do so.

+++ src/h1_htx.c
@@ -477,7 +477,7 @@ int h1_parse_msg_data(struct h1m *h1m, struct htx **dsthtx,
 			total += ret;
 		}
 		if (h1m->state == H1_MSG_CHUNK_SIZE) {
-			unsigned int chksz;
+			uint64_t chksz;
 
 			ret = h1_parse_chunk_size(srcbuf, ofs, b_data(srcbuf), &chksz);
 			if (ret <= 0)
